+++ fs/djs/module.f.ts
@@ -3,7 +3,7 @@
  *
  * @module
  */
-import type { Io } from '../io/module.f.ts'
+import { fromIo, type Io } from '../io/module.f.ts'
 import type { Primitive as JsonPrimitive } from '../json/module.f.ts'
 import { transpile } from './transpiler/module.f.ts'
 import { stringify, stringifyAsTree } from './serializer/module.f.ts'
@@ -20,15 +20,16 @@ export type Primitive = JsonPrimitive | bigint | undefined
 
 export type Unknown = Primitive | Object | Array
 
-export const compile = ({ console: { error }, fs }: Io) => (args: readonly string[]): Promise<number> => {
+export const compile = (io: Io) => async (args: readonly string[]): Promise<number> => {
+    const { console: { error: logError }, fs } = io
     if (args.length < 2) {
-        error('Error: Requires 2 or more arguments')
-        return Promise.resolve(1)
+        logError('Error: Requires 2 or more arguments')
+        return 1
     }
 
     const inputFileName = args[0]
     const outputFileName = args[1]
-    const result = transpile(fs)(inputFileName)
+    const result = await fromIo(io)(transpile(inputFileName))
     switch (result[0]) {
         case 'ok': {
             if (outputFileName.endsWith('.json'))
@@ -43,9 +44,9 @@ export const compile = ({ console: { error }, fs }: Io) => (args: readonly strin
         }
         case 'error': {
             const metadata = result[1].metadata
-            error(`${metadata?.path}:${metadata?.line}:${metadata?.column} - error: ${result[1].message}`)
+            logError(`${metadata?.path}:${metadata?.line}:${metadata?.column} - error: ${result[1].message}`)
             break
         }
     }
-    return Promise.resolve(0)
+    return 0
 }